Start to split code into components
